Streamlined the scheduler and renamed tasks
The schedule is now much simpler! Also renamed tasks with "cleanup" in the name to "purge" and vice versa to accomodate the new schedule. 
NEW: make sure you add the command to your torrent client to process files immediately after download is finished (your library will be updated as well).